articles/how-to-implement-stable-diffusion/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision

transformers
accelerate
diffusers
tqdm
pillow
